update proj.android for cocos2dx 3.17
