--- NinjaFightMultiplayerDemo/proj.android/.settings/org.eclipse.cdt.core.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-environment/project/0.1230402123/append=true
-environment/project/0.1230402123/appendContributed=true
