+++ NinjaFightMultiplayerDemo/proj.android/app/.gitignore
@@ -1,0 +1,2 @@
+/build
+/.externalNativeBuild
