+++ NinjaFightMultiplayerDemo/proj.android/app/jni/Application.mk
@@ -1,0 +1,16 @@
+APP_STL := c++_static
+
+APP_CPPFLAGS := -frtti -DCC_ENABLE_CHIPMUNK_INTEGRATION=1 -std=c++11 -fsigned-char -Wno-extern-c-compat
+APP_LDFLAGS := -latomic
+
+APP_ABI := armeabi-v7a
+APP_SHORT_COMMANDS := true
+
+
+ifeq ($(NDK_DEBUG),1)
+  APP_CPPFLAGS += -DCOCOS2D_DEBUG=1
+  APP_OPTIM := debug
+else
+  APP_CPPFLAGS += -DNDEBUG
+  APP_OPTIM := release
+endif
