--- NinjaFightMultiplayerDemo/proj.android/build-cfg.json
@@ -1,13 +1,0 @@
-{
-    "ndk_module_path" :[
-        "../cocos2d",
-        "../cocos2d/cocos",
-        "../cocos2d/external"
-    ],
-    "copy_resources": [
-        {
-            "from": "../Resources",
-            "to": ""
-        }
-    ]
-}
