+++ NinjaFightMultiplayerDemo/proj.android/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Tue Mar 14 17:40:59 CST 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.1-all.zip
