--- NinjaFightMultiplayerDemo/proj.android/jni/Android.mk
@@ -1,47 +1,0 @@
-LOCAL_PATH := $(call my-dir)
-
-include $(CLEAR_VARS)
-
-$(call import-add-path,$(LOCAL_PATH)/../../cocos2d)
-$(call import-add-path,$(LOCAL_PATH)/../../cocos2d/external)
-$(call import-add-path,$(LOCAL_PATH)/../../cocos2d/cocos)
-
-LOCAL_MODULE := cocos2dcpp_shared
-
-LOCAL_MODULE_FILENAME := libcocos2dcpp
-
-LOCAL_SRC_FILES := hellocpp/main.cpp \
-                   ../../Classes/AppDelegate.cpp \
-                   ../../Classes/HelloWorldScene.cpp \
-                   ../../Classes/Player.cpp \
-                   ../../Classes/StartGameLayer.cpp \
-                   ../../AppWarpX/HMAC_SHA1.cpp \
-                   ../../AppWarpX/SHA1.cpp \
-                   ../../AppWarpX/appwarp.cpp \
-                   ../../AppWarpX/appwarp_extended.cpp \
-                   ../../AppWarpX/base64.cpp \
-                   ../../AppWarpX/requests.cpp \
-                   ../../AppWarpX/socket.cpp \
-                   ../../AppWarpX/udpsocket.cpp \
-                   ../../AppWarpX/urlencode.cpp \
-                   ../../AppWarpX/utilities.cpp \
-                   ../../AppWarpX/cJSON.c
-
-LOCAL_C_INCLUDES := $(LOCAL_PATH)/../../Classes \
-                    $(LOCAL_PATH)/../../AppWarpX
-
-# _COCOS_HEADER_ANDROID_BEGIN
-# _COCOS_HEADER_ANDROID_END
-
-
-LOCAL_STATIC_LIBRARIES := cocos2dx_static
-
-# _COCOS_LIB_ANDROID_BEGIN
-# _COCOS_LIB_ANDROID_END
-
-include $(BUILD_SHARED_LIBRARY)
-
-$(call import-module,.)
-
-# _COCOS_LIB_IMPORT_ANDROID_BEGIN
-# _COCOS_LIB_IMPORT_ANDROID_END
