--- NinjaFightMultiplayerDemo/proj.android/project.properties
@@ -1,13 +1,0 @@
-# This file is automatically generated by Android Tools.
-# Do not modify this file -- YOUR CHANGES WILL BE ERASED!
-#
-# This file must be checked in Version Control Systems.
-#
-# To customize properties used by the Ant build system use,
-# "ant.properties", and override values to adapt the script to your
-# project structure.
-
-# Project target.
-target=android-10
-
-android.library.reference.1=../cocos2d/cocos/platform/android/java
