+++ NinjaFightMultiplayerDemo/proj.android/settings.gradle
@@ -1,0 +1,4 @@
+include ':libcocos2dx'
+project(':libcocos2dx').projectDir = new File(settingsDir, '../cocos2d/cocos/platform/android/libcocos2dx')
+include ':NinjaFight'
+project(':NinjaFight').projectDir = new File(settingsDir, 'app')
